EMIS/requirements.txt
django
djangorestframework
django-rest-framework-simplejwt
psycopg